+++ ivy/functional/backends/torch/experimental/statistical.py
@@ -122,7 +122,7 @@ def unravel_index(
     *,
     out: Optional[torch.Tensor] = None,
 ) -> Tuple:
-    temp = indices.to(torch.int32)
+    temp = indices
     output = []
     for dim in reversed(shape):
         output.append(temp % dim)
